legalAI-backend/requirements.txt
fastapi==0.74.0
openai==0.11.0
python-dotenv==0.19.2
uvicorn==0.17.5
bidict==0.21.3
click==7.1.2
fastapi-socketio==0.0.8
h11==0.9.0
httptools==0.1.2
pydantic==1.8.2
python-engineio==4.2.1
python-socketio==5.4.0
starlette==0.14.2
typing-extensions==3.10.0.2
uvloop==0.16.0
websockets==8.1